style: modify the unit test structure
