+++ test/logic/validator/Number.test.js
@@ -1,0 +1,183 @@
+import { describe, assert, expect, test } from 'vitest'
+import { validator } from '@ER/hooks/use-logic'
+describe('validator', () => {
+  test('Number', () => {
+    const field = { type: 'number', label: 'Number', icon: 'numbers', key: 'number_3bdLUMFSVB7_YujfBJhBB', id: '3bdLUMFSVB7_YujfBJhBB', options: { min: 0, max: null, step: 1, precision: 0, disabled: false, controls: true, controlsPosition: true, defaultValue: null, labelWidth: 100, isShowLabel: true, required: false, minlength: null, maxlength: null }, style: { width: { pc: '100%', mobile: '100%' } } }
+    expect(
+      validator({
+        operator: 'equal',
+        value: 123
+      },
+      123.00,
+      field)).toBeTruthy()
+    expect(
+      validator({
+        operator: 'not_equal',
+        value: 123
+      },
+      123.00,
+      field)).toBeFalsy()
+    expect(
+      validator({
+        operator: 'empty'
+      },
+      '',
+      field)).toBeTruthy()
+    expect(
+      validator({
+        operator: 'empty'
+      },
+      undefined,
+      field)).toBeTruthy()
+    expect(
+      validator({
+        operator: 'empty'
+      },
+      null,
+      field)).toBeTruthy()
+    expect(
+      validator({
+        operator: 'not_empty'
+      },
+      '',
+      field)).toBeFalsy()
+    expect(
+      validator({
+        operator: 'not_empty'
+      },
+      undefined,
+      field)).toBeFalsy()
+    expect(
+      validator({
+        operator: 'not_empty'
+      },
+      null,
+      field)).toBeFalsy()
+    expect(
+      validator({
+        operator: 'greater_than',
+        value: 2
+      },
+      null,
+      field)).toBeFalsy()
+    expect(
+      validator({
+        operator: 'greater_than',
+        value: 2
+      },
+      '',
+      field)).toBeFalsy()
+    expect(
+      validator({
+        operator: 'greater_than',
+        value: 2
+      },
+      undefined,
+      field)).toBeFalsy()
+    expect(
+      validator({
+        operator: 'greater_than',
+        value: 2
+      },
+      2.0,
+      field)).toBeFalsy()
+    expect(
+      validator({
+        operator: 'greater_than',
+        value: 2
+      },
+      2.1,
+      field)).toBeTruthy()
+    expect(
+      validator({
+        operator: 'greater_than_equal',
+        value: 2
+      },
+      2.1,
+      field)).toBeTruthy()
+    expect(
+      validator({
+        operator: 'less_than',
+        value: 2
+      },
+      null,
+      field)).toBeTruthy()
+    expect(
+      validator({
+        operator: 'less_than',
+        value: 2
+      },
+      '',
+      field)).toBeTruthy()
+    expect(
+      validator({
+        operator: 'less_than',
+        value: 2
+      },
+      undefined,
+      field)).toBeTruthy()
+    expect(
+      validator({
+        operator: 'less_than',
+        value: 2
+      },
+      2.0,
+      field)).toBeFalsy()
+    expect(
+      validator({
+        operator: 'less_than',
+        value: 2
+      },
+      2.1,
+      field)).toBeFalsy()
+    expect(
+      validator({
+        operator: 'less_than_equal',
+        value: 2
+      },
+      2.0,
+      field)).toBeTruthy()
+    expect(
+      validator({
+        operator: 'less_than_equal',
+        value: 2
+      },
+      1.9,
+      field)).toBeTruthy()
+    expect(
+      validator({
+        operator: 'between',
+        value: [1, 2]
+      },
+      1,
+      field)).toBeTruthy()
+    expect(
+      validator({
+        operator: 'between',
+        value: [1, 2]
+      },
+      2.0,
+      field)).toBeTruthy()
+    expect(
+      validator({
+        operator: 'between',
+        value: [1, 2]
+      },
+      null,
+      field)).toBeFalsy()
+    expect(
+      validator({
+        operator: 'between',
+        value: [1, 2]
+      },
+      '',
+      field)).toBeFalsy()
+    expect(
+      validator({
+        operator: 'between',
+        value: [1, 2]
+      },
+      undefined,
+      field)).toBeFalsy()
+  })
+})
