+++ test/logic/validator/Textarea.test.js
@@ -1,0 +1,75 @@
+import { describe, assert, expect, test } from 'vitest'
+import { validator } from '@ER/hooks/use-logic'
+describe('validator', () => {
+  test('Textarea', () => {
+    const field = { type: 'textarea', label: 'Textarea', icon: 'textarea', key: 'textarea_ZPRzSVZkMHOMYgWZBmrvR', id: 'ZPRzSVZkMHOMYgWZBmrvR', options: { clearable: true, isShowWordLimit: false, rows: 6, defaultValue: '', placeholder: 'Please enter', disabled: false, labelWidth: 100, isShowLabel: true, required: false, min: null, max: null }, style: { width: { pc: '100%', mobile: '100%' } } }
+    expect(
+      validator({
+        operator: 'equal',
+        value: '123'
+      },
+      '123',
+      field)).toBeTruthy()
+    expect(
+      validator({
+        operator: 'not_equal',
+        value: '123'
+      },
+      '123',
+      field)).toBeFalsy()
+    expect(
+      validator({
+        operator: 'contains',
+        value: [
+          '123'
+        ]
+      },
+      '123',
+      field)).toBeTruthy()
+    expect(
+      validator({
+        operator: 'not_contain',
+        value: [
+          '123'
+        ]
+      },
+      '123',
+      field)).toBeFalsy()
+    expect(
+      validator({
+        operator: 'empty'
+      },
+      '',
+      field)).toBeTruthy()
+    expect(
+      validator({
+        operator: 'empty'
+      },
+      undefined,
+      field)).toBeTruthy()
+    expect(
+      validator({
+        operator: 'empty'
+      },
+      null,
+      field)).toBeTruthy()
+    expect(
+      validator({
+        operator: 'not_empty'
+      },
+      '',
+      field)).toBeFalsy()
+    expect(
+      validator({
+        operator: 'not_empty'
+      },
+      undefined,
+      field)).toBeFalsy()
+    expect(
+      validator({
+        operator: 'not_empty'
+      },
+      null,
+      field)).toBeFalsy()
+  })
+})
